working after talk with Anastasis
